+++ CHANGELOG.md
@@ -13,10 +13,11 @@ Documentation here: <https://py-maybetype.readthedocs.io/en/latest/reference/>
 
 ### Added
 
-- Added the `result` module
-  - Added classes `Result`, `Ok`, and `Err`
+- Added classes `Result`, `Ok`, and `Err` to mimic the behavior of Rust's `Result` type
+  - See docs for all introduced methods: <https://py-maybetype.readthedocs.io/en/latest/reference/index.html#maybetype.Result>
 - Added the `errors` module
   - Added exception `ResultUnwrapError`
+- Added test module `test_result`
 
 ### Changed
 
@@ -36,7 +37,7 @@ Documentation here: <https://py-maybetype.readthedocs.io/en/latest/reference/>
 ### Added
 
 - Added method `Maybe.cast`
-  - Casts the type of the wrapped value: `maybe(...).cast(int)` → `maybe[int]`
+  - Casts the type of the wrapped value: `maybe(...).cast(int)` → `Maybe[int]`
 - Added method `Maybe.flatten`
   - Replicates Rust's [`Option::flatten`](https://doc.rust-lang.org/std/option/enum.Option.html#method.flatten)
 - Added method `Maybe.inspect`