+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Generally, more significant breaking changes will be put near the top of each ca
 
 Documentation here: <https://py-maybetype.readthedocs.io/en/latest/reference/>
 
-## Unreleased
+## [0.11.0] - 2026-04-06
 
 ### Added